% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eif}
\alias{eif}
\title{Simulated Efficient Influence Functions}
\format{
An object of class \code{list} of length 5.
}
\usage{
eif
}
\description{
A list of 5 simulated influence functions.
}
\keyword{datasets}
